+++ packages/dart_pdf_editor/lib/l10n/dart_pdf_editor_localizations_nl.dart
@@ -2140,4 +2140,8 @@ class DartPdfEditorLocalizationsNl extends DartPdfEditorLocalizations {
 
   @override
   String get splitFailed => 'Deze PDF kon niet worden gesplitst.';
+
+  @override
+  String get guidesSnapHint =>
+      'Randen en middelpunten van annotaties uitlijnen • Houd Alt ingedrukt om over te slaan';
 }